requirements/base.txt
django==4.0.2
django-environ==0.8.1
django-model-utils==4.2.0
psycopg2==2.9.3